refactor(settings): split account-section into profile, passkey, and avatar modules
